reorganising components and files
